refactor: RacingGameView class를 RacingGameInputView, RacingGameOutputView로 나눔
